Convert to ESP32-S3
My ESP32 does not support 2,000,000 baud on the Serial Port, this will be an issue for supporting CanDrive
Also fix the SavvyCAN header format.
NOTE: For bench testing it is important to connect the power pack to the circuit so the OBD2_Interface_v2 circuit is powered via the OBD2 port as it would be if attached to the car.
Car Tested and working fine